Se completaron los repositorios y avances en ClienteService

+++ src/main/java/com/comercio/app/services/IBaseService.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface IBaseService<E> {
     public List<E> findAll() throws Exception;
-    public E findById(Long id);
+    public E findById(Integer id);
     public E insert(E entity);
-    public E update(Long id, E entity);
-    public boolean delete(Long id);
+    public E update(Integer id, E entity);
+    public boolean delete(Integer id);
 }
